implementations/uimServicemanager/agent-query-service/requirements.txt
# Core dependencies
faststream[nats]>=0.5.0
pydantic>=2.0.0
pydantic-ai>=0.0.14

# HTTP client for calling catalogue API
httpx>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Optional: for local testing without OpenAI
# ollama  # Uncomment if you want to use local LLMs
